Reorganize project
server and client seperated.
Tests live beside the code they test.
