+++ package.json
@@ -7,7 +7,8 @@
     "postinstall": "npm run build",
     "server": "node start.js",
     "build": "webpack -p --progress --config webpack.production.config.js",
-    "test": "nyc ava --fail-fast --verbose --require babel-register",
+    "test": "nyc ava",
+    "test-watch": "nyc ava --watch",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
     "clean": "rm build/app.js",
     "lint": "eslint"
@@ -28,11 +29,24 @@
     "url": "https://github.com/pheuter/essential-react/issues"
   },
   "homepage": "https://github.com/pheuter/essential-react",
+  "ava": {
+    "verbose": true,
+    "failFast": true,
+    "files": ["src/**/__tests__/**/*.js"],
+    "source": [
+      "src/**/*",
+      "!build/**/*",
+      "!build_support/**/*"
+    ],
+    "require": ["babel-register"],
+    "babel": "inherit"
+  },
   "devDependencies": {
     "autoprefixer": "^6.0.2",
     "ava": "^0.12.0",
     "babel-eslint": "6.1.2",
     "babel-loader": "^6.2.4",
+    "babel-plugin-resolver": "0.0.7",
     "babel-plugin-transform-runtime": "^6.5.2",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
