+++ src/client/assets/static/images/apple-logo.png
