+++ src/client/common/components/App.js
