+++ src/client/common/components/Routes.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { Route, IndexRoute } from 'react-router'
 
 import App from './App'
-import LoginPage from '../../pages/login/page'
-import HomePage from '../../pages/home/page'
-import PingPage from '../../pages/ping/page'
-import Terminal from '../../pages/terminal'
+import LoginPage from 'client/pages/login/page'
+import HomePage from 'client/pages/home/page'
+import PingPage from 'client/pages/ping/page'
+import Terminal from 'client/pages/terminal'
 
 
 export default (
