+++ src/client/common/components/__tests__/App-test.js
@@ -2,8 +2,7 @@ import test from 'ava'
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import App from '../../../src/common/components/App'
-
+import App from '../App'
 
 test('render with container div', t => {
   const wrapper = shallow(React.createElement(App))
