+++ src/client/pages/ping/page.js
