+++ src/client/pages/terminal/Components/KeyManager.js
