+++ src/client/pages/terminal/index.js
