+++ src/client/pages/terminal/page.css
