+++ src/server/api/loginTerminal/commands/cat.js
