+++ webpack.production.config.js
@@ -5,7 +5,7 @@ var ExtractTextPlugin = require('extract-text-webpack-plugin');
  * This is the Webpack configuration file for production.
  */
 module.exports = {
-  entry: "./src/main",
+  entry: "./src/client/main",
 
   output: {
     path: __dirname + "/build/",
@@ -24,7 +24,6 @@ module.exports = {
   },
 
   resolve: {
-    modulesDirectories: ['node_modules', 'src'],
     extensions: ['', '.js', '.jsx', '.css']
   },
